Fix naming of form fields and elements folders
