+++ src/components/shared/form_elements/RadioFormInput.vue
@@ -22,8 +22,8 @@
 
 <script setup lang="ts">
 
-import { useInputFocusing } from '@src/components/shared/form_inputs/useInputFocusing';
-import { useInputModel } from '@src/components/shared/form_inputs/useInputModel';
+import { useInputFocusing } from '@src/components/shared/form_elements/useInputFocusing';
+import { useInputModel } from '@src/components/shared/form_elements/useInputModel';
 
 interface Props {
 	modelValue: string | number | boolean | null;
