+++ src/components/shared/form_elements/SelectFormInput.vue
@@ -16,7 +16,7 @@
 
 <script setup lang="ts">
 
-import { useInputModel } from '@src/components/shared/form_inputs/useInputModel';
+import { useInputModel } from '@src/components/shared/form_elements/useInputModel';
 
 interface Props {
 	modelValue: string | number;
