+++ src/components/shared/form_fields/CountryAutocompleteField.vue
@@ -37,8 +37,8 @@
 import { useCountryInput } from '@src/components/shared/legacy_form_inputs/useCountryInput';
 import { useFilteredCountries } from '@src/components/shared/legacy_form_inputs/useFilteredCountries';
 import { Country } from '@src/view_models/Country';
-import TextFormInput from '@src/components/shared/form_inputs/TextFormInput.vue';
-import { useCountryAutocompleteEvents } from '@src/components/shared/form_elements/useCountryAutocompleteEvents';
+import TextFormInput from '@src/components/shared/form_elements/TextFormInput.vue';
+import { useCountryAutocompleteEvents } from '@src/components/shared/form_fields/useCountryAutocompleteEvents';
 
 interface Props {
 	modelValue: string;
