+++ src/components/shared/form_fields/EmailField.vue
@@ -22,10 +22,10 @@
 </template>
 
 <script setup lang="ts">
-import { useFieldModel } from '@src/components/shared/form_elements/useFieldModel';
-import { useSuggestedEmailProvider } from '@src/components/shared/form_elements/useSuggestedEmailProvider';
+import { useFieldModel } from '@src/components/shared/form_fields/useFieldModel';
+import { useSuggestedEmailProvider } from '@src/components/shared/form_fields/useSuggestedEmailProvider';
 import { useMailHostList } from '@src/components/shared/useMailHostList';
-import TextFormInput from '@src/components/shared/form_inputs/TextFormInput.vue';
+import TextFormInput from '@src/components/shared/form_elements/TextFormInput.vue';
 
 interface Props {
 	modelValue: string;
