+++ src/components/shared/form_fields/RadioField.vue
@@ -23,9 +23,9 @@
 </template>
 
 <script setup lang="ts">
-import { FormOption } from '@src/components/shared/form_elements/FormOption';
-import RadioFormInput from '@src/components/shared/form_inputs/RadioFormInput.vue';
-import { useFieldModel } from '@src/components/shared/form_elements/useFieldModel';
+import { FormOption } from '@src/components/shared/form_fields/FormOption';
+import RadioFormInput from '@src/components/shared/form_elements/RadioFormInput.vue';
+import { useFieldModel } from '@src/components/shared/form_fields/useFieldModel';
 
 interface Props {
 	label: String;
