+++ src/components/shared/form_fields/SelectField.vue
@@ -16,9 +16,9 @@
 
 <script setup lang="ts">
 
-import { FormOption } from '@src/components/shared/form_elements/FormOption';
-import { useFieldModel } from '@src/components/shared/form_elements/useFieldModel';
-import SelectFormInput from '@src/components/shared/form_inputs/SelectFormInput.vue';
+import { FormOption } from '@src/components/shared/form_fields/FormOption';
+import { useFieldModel } from '@src/components/shared/form_fields/useFieldModel';
+import SelectFormInput from '@src/components/shared/form_elements/SelectFormInput.vue';
 
 interface Props {
 	label: String;
