--- src/scss/components/button.scss
@@ -1,3 +1,0 @@
-.form-button {
-
-}
